RockPaperScissorsGame
Created the rock paper scissors game in node environment and also installed express for future usage. The code uses Math.Random() function to generate a random option for the computer side and the user playing the game selects their option. Then the code defines whether the user wins or loses that particular round.