Created the portfolio website
Included all the necessary files like resume and also included github, linkedin and facebook link in the contact me section. Designed the UI and included the github links of the projects made till date. Utilized a bit of the bootstrap for styling not much though. Made sure that the project runs from index.js.